Sync procs limits for csf_flood_guard and second_flood_guard

+++ aegir/tools/system/guest-fire.sh
@@ -5,17 +5,17 @@ SHELL=/bin/bash
 
 second_flood_guard() {
   thisCountSec=`ps aux | grep -v "grep" | grep -v "null" | grep --count "second.sh"`
-  if [ ${thisCountSec} -gt "2" ]; then
+  if [ ${thisCountSec} -gt "4" ]; then
     echo "$(date 2>&1) Too many ${thisCountSec} second.sh processes killed" >> \
       /var/log/sec-count.kill.log
     kill -9 $(ps aux | grep '[s]econd.sh' | awk '{print $2}') &> /dev/null
   fi
 }
-second_flood_guard
+[ ! -e "/var/run/boa_run.pid" ] && second_flood_guard
 
 csf_flood_guard() {
   thisCountCsf=`ps aux | grep -v "grep" | grep -v "null" | grep --count "csf"`
-  if [ ${thisCountCsf} -gt "3" ]; then
+  if [ ! -e "/var/run/boa_run.pid" ] && [ ${thisCountCsf} -gt "4" ]; then
     echo "$(date 2>&1) Too many ${thisCountCsf} csf processes killed" >> \
       /var/log/csf-count.kill.log
     kill -9 $(ps aux | grep '[c]sf' | awk '{print $2}') &> /dev/null
@@ -23,13 +23,13 @@ csf_flood_guard() {
     csf -df
   fi
   thisCountFire=`ps aux | grep -v "grep" | grep -v "null" | grep --count "fire.sh"`
-  if [ ${thisCountFire} -gt "9" ]; then
+  if [ ! -e "/var/run/boa_run.pid" ] && [ ${thisCountFire} -gt "9" ]; then
     echo "$(date 2>&1) Too many ${thisCountFire} fire.sh processes killed and rules purged" >> \
       /var/log/fire-purge.kill.log
     csf -tf
     csf -df
     kill -9 $(ps aux | grep '[f]ire.sh' | awk '{print $2}') &> /dev/null
-  elif [ ${thisCountFire} -gt "7" ]; then
+  elif [ ! -e "/var/run/boa_run.pid" ] && [ ${thisCountFire} -gt "7" ]; then
     echo "$(date 2>&1) Too many ${thisCountFire} fire.sh processes killed" >> \
       /var/log/fire-count.kill.log
     csf -tf
